poc/requirements.txt
google-auth-oauthlib
google-api-python-client
google-genai
PyPDF2
python-docx
openpyxl
python-pptx
python-dotenv
tenacity
pyyaml
langchain
langchain-google-genai
reportlab  # Office形式→PDF変換用（オプション）
tqdm  # 進捗表示用
unstructured[pdf,md,docx,xlsx,pptx]
langchain-text-splitters  # テキスト分割用
func-timeout  # タイムアウト処理用
